add basic folder structure, fix formgroup, moved some files, update file names
