+++ src/app/app.module.ts
@@ -3,8 +3,8 @@ import { NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import {HttpClientModule} from '@angular/common/http';
-import { CategoryModule } from './category/category.module';
+import { HttpClientModule } from '@angular/common/http';
+import { CategoryModule } from './modules/category.module';
 
 @NgModule({
   declarations: [
