+++ src/app/components/category/category-edit/category-edit.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
-import {CategoryService} from './category-service';
-import { Category } from './category';
+import {CategoryService} from '../../../services/category-service';
+import { Category } from '../../../models/category.model';
 import {FormGroup, FormBuilder, Validators} from '@angular/forms';
 
 @Component({
