+++ src/app/components/category/category-list/category-list.component.html
