+++ src/app/components/category/category-list/category-list.component.spec.ts
@@ -1,4 +1,7 @@
+/* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+import { DebugElement } from '@angular/core';
 
 import { CategoryListComponent } from './category-list.component';
