+++ src/app/components/category/category-list/category-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import {CategoryService} from './category-service';
-import { Category } from './category';
+import { CategoryService } from '../../../services/category-service';
+import { Category } from '../../../models/category.model';
 
 @Component({
   templateUrl: './category-list.component.html',
