+++ src/app/modules/category.module.ts
@@ -1,15 +1,16 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {CategoryListComponent} from './category-list.component';
-import {CategoryEditComponent} from './category-edit.component';
-import {RouterModule} from '@angular/router';
-import {ReactiveFormsModule} from '@angular/forms';
+import { CategoryListComponent } from '../components/category/category-list/category-list.component';
+import { CategoryEditComponent } from '../components/category/category-edit/category-edit.component';
+import { RouterModule } from '@angular/router';
+import { ReactiveFormsModule } from '@angular/forms';
 import { InMemoryWebApiModule } from 'angular-in-memory-web-api';
-import {CategoryData} from './category-data';
+import { CategoryData } from '../data/category.data';
 
 @NgModule({
   declarations: [
-    CategoryListComponent
+    CategoryListComponent,
+    CategoryEditComponent
   ],
   imports: [
     CommonModule,
